pipe

+++ angular/src/app/components/song-details/song-details.component.ts
@@ -11,26 +11,44 @@ import {AuthService} from "../../services/auth.service";
 })
 export class SongDetailsComponent {
   song : Song = { ...defaultSong };
+  user : string = '';
+  showFullText = false;
 
   constructor(private service: DataService, private authService: AuthService, private route: ActivatedRoute, public router: Router) {}
 
   ngOnInit() {
-    this.getById();
+    this.getData();
   }
 
-  getById() {
+  getData() {
     this.route.queryParams.subscribe(params => {
       const id = params['songId'];
       this.service.getById(id).subscribe(response => {
         this.song = response as Song;
+
+        if(this.authService.isLoggedIn() && this.song.userId === this.authService.getUserId()) {
+          this.user = "Zamieściłeś ty";
+        } else {
+          this.service.getUserName(this.song.userId).subscribe(response => {
+           this.user = "Zamieścił: " + response as string;
+          });
+        }
       });
     });
   }
 
   delete() {
-    this.service.deleteMusic(this.song.id as string).subscribe(response => {
-      console.log(response);
-    });
-    this.router.navigate(['/']);
+    if(this.authService.isLoggedIn() && this.authService.getUserId() === this.song.userId) {
+      this.service.deleteMusic(this.song.id as string).subscribe(response => {
+        console.log(response);
+      });
+      this.router.navigate(['/']);
+
+    }
   }
+
+  toggleFullText() {
+    this.showFullText = !this.showFullText;
+  }
+
 }
